[#1745] Rename generic pointer into PointerFamily
