+++ iceoryx2-bb/container/src/slotmap.rs
@@ -54,9 +54,9 @@ use iceoryx2_bb_concurrency::atomic::AtomicBool;
 use iceoryx2_bb_derive_macros::ZeroCopySend;
 use iceoryx2_bb_elementary::bump_allocator::BumpAllocator;
 use iceoryx2_bb_elementary::relocatable_ptr::GenericRelocatablePointer;
-use iceoryx2_bb_elementary_traits::generic_pointer::{NonNullFamily, PointerFamily};
 use iceoryx2_bb_elementary_traits::owning_pointer::GenericOwningPointer;
 use iceoryx2_bb_elementary_traits::placement_default::PlacementDefault;
+use iceoryx2_bb_elementary_traits::pointer_family::{NonNullFamily, PointerFamily};
 pub use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
 use iceoryx2_bb_elementary_traits::testing::abandonable::Abandonable;
 use iceoryx2_bb_elementary_traits::zero_copy_send::ZeroCopySend;
