+++ iceoryx2-bb/container/src/string/relocatable_string.rs
@@ -83,8 +83,8 @@ use core::mem::MaybeUninit;
 use core::ops::{Deref, DerefMut};
 use iceoryx2_bb_elementary::math::unaligned_mem_size;
 use iceoryx2_bb_elementary::relocatable_ptr::RelocatablePointer;
-use iceoryx2_bb_elementary_traits::generic_pointer::NonNullFamily;
 use iceoryx2_bb_elementary_traits::pointer::Pointer;
+use iceoryx2_bb_elementary_traits::pointer_family::NonNullFamily;
 pub use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
 use iceoryx2_log::{fail, fatal_panic};
