+++ iceoryx2-bb/container/src/vector/polymorphic_vec.rs
@@ -65,8 +65,8 @@ use core::{
 
 use iceoryx2_bb_elementary_traits::{
     allocator::{AllocationError, BaseAllocator},
-    generic_pointer::NonNullFamily,
     pointer::Pointer,
+    pointer_family::NonNullFamily,
 };
 use iceoryx2_log::fail;
