+++ iceoryx2-bb/container/src/vector/relocatable_vec.rs
@@ -80,7 +80,7 @@ use core::{fmt::Debug, mem::MaybeUninit};
 
 use iceoryx2_bb_elementary::{math::unaligned_mem_size, relocatable_ptr::*};
 pub use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
-use iceoryx2_bb_elementary_traits::{generic_pointer::NonNullFamily, zero_copy_send::ZeroCopySend};
+use iceoryx2_bb_elementary_traits::{pointer_family::NonNullFamily, zero_copy_send::ZeroCopySend};
 use iceoryx2_log::{fail, fatal_panic};
 
 pub use crate::vector::Vector;
