+++ iceoryx2-bb/elementary-traits/src/lib.rs
@@ -21,12 +21,12 @@ pub mod allocator;
 mod as_cstr;
 pub use crate::as_cstr::*;
 pub mod atomic_copy;
-pub mod generic_pointer;
 pub mod iceoryx_send;
 pub mod owning_pointer;
 pub mod placement_default;
 pub mod plain_old_data_without_padding;
 pub mod pointer;
+pub mod pointer_family;
 pub mod relocatable_container;
 pub mod testing;
 pub mod type_name;
