+++ iceoryx2-bb/elementary/src/relocatable_ptr.rs
@@ -68,7 +68,7 @@ use core::{fmt::Debug, marker::PhantomData, ptr::NonNull};
 
 pub use iceoryx2_bb_elementary_traits::pointer::Pointer;
 
-use iceoryx2_bb_elementary_traits::{generic_pointer::PointerFamily, zero_copy_send::ZeroCopySend};
+use iceoryx2_bb_elementary_traits::{pointer_family::PointerFamily, zero_copy_send::ZeroCopySend};
 use iceoryx2_pal_concurrency_sync::atomic::{AtomicIsize, Ordering};
 
 #[derive(Debug)]
