+++ iceoryx2-bb/lock-free/src/mpmc/bit_set.rs
@@ -62,7 +62,7 @@ pub type RelocatableBitSet = details::BitSet<RelocatablePointer<details::BitsetE
 pub mod details {
 
     use iceoryx2_bb_elementary_traits::{
-        generic_pointer::NonNullFamily, zero_copy_send::ZeroCopySend,
+        pointer_family::NonNullFamily, zero_copy_send::ZeroCopySend,
     };
 
     use super::*;
