+++ iceoryx2-bb/lock-free/src/mpmc/counting_bit_set.rs
@@ -107,7 +107,7 @@ pub type RelocatableCountingBitSet = details::CountingBitSet<RelocatablePointer<
 
 #[doc(hidden)]
 pub mod details {
-    use iceoryx2_bb_elementary_traits::generic_pointer::NonNullFamily;
+    use iceoryx2_bb_elementary_traits::pointer_family::NonNullFamily;
 
     use super::*;
