+++ iceoryx2-bb/lock-free/src/spsc/index_queue.rs
@@ -106,7 +106,7 @@ pub type RelocatableIndexQueue = details::IndexQueue<RelocatablePointer<UnsafeCe
 
 pub mod details {
     use iceoryx2_bb_elementary_traits::{
-        generic_pointer::NonNullFamily, zero_copy_send::ZeroCopySend,
+        pointer_family::NonNullFamily, zero_copy_send::ZeroCopySend,
     };
 
     use super::*;
