+++ iceoryx2-bb/memory/src/heap_allocator.rs
@@ -16,7 +16,7 @@ use core::{alloc::Layout, ptr::NonNull};
 
 use iceoryx2_bb_elementary_traits::{
     allocator::{AllocationGrowError, AllocationShrinkError, ReallocGrow, ReallocShrink},
-    generic_pointer::NonNullFamily,
+    pointer_family::NonNullFamily,
 };
 use iceoryx2_bb_posix::memory::heap;
 use iceoryx2_log::fail;
