+++ iceoryx2-bb/memory/src/pool_allocator.rs
@@ -57,8 +57,8 @@ use iceoryx2_bb_elementary::bump_allocator::BumpAllocator;
 use iceoryx2_bb_elementary::math::align;
 use iceoryx2_bb_elementary::sync_ptr::SyncPtr;
 pub use iceoryx2_bb_elementary_traits::allocator::*;
-use iceoryx2_bb_elementary_traits::generic_pointer::NonNullFamily;
 use iceoryx2_bb_elementary_traits::pointer::Pointer;
+use iceoryx2_bb_elementary_traits::pointer_family::NonNullFamily;
 use iceoryx2_bb_elementary_traits::relocatable_container::*;
 use iceoryx2_bb_elementary_traits::zero_copy_send::ZeroCopySend;
 use iceoryx2_bb_lock_free::mpmc::unique_index_set::*;
