+++ iceoryx2-cal/src/zero_copy_connection/common.rs
@@ -24,7 +24,7 @@ pub mod details {
     use iceoryx2_bb_concurrency::cell::UnsafeCell;
     use iceoryx2_bb_container::vector::relocatable_vec::*;
     use iceoryx2_bb_elementary_traits::allocator::{AllocationError, BaseAllocator};
-    use iceoryx2_bb_elementary_traits::generic_pointer::NonNullFamily;
+    use iceoryx2_bb_elementary_traits::pointer_family::NonNullFamily;
     use iceoryx2_bb_elementary_traits::relocatable_container::RelocatableContainer;
     use iceoryx2_bb_lock_free::spsc::{
         index_queue::RelocatableIndexQueue,
