+++ iceoryx2-cal/src/zero_copy_connection/used_chunk_list.rs
@@ -31,7 +31,7 @@ pub mod details {
 
     use iceoryx2_bb_elementary::math::unaligned_mem_size;
     use iceoryx2_bb_elementary_traits::{
-        generic_pointer::NonNullFamily, owning_pointer::OwningPointer, zero_copy_send::ZeroCopySend,
+        pointer_family::NonNullFamily, owning_pointer::OwningPointer, zero_copy_send::ZeroCopySend,
     };
 
     use super::*;
